+++ plugins/scaffolder-backend-module-github/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@backstage/plugin-scaffolder-backend-module-github",
+  "description": "The github module for @backstage/plugin-scaffolder-backend",
+  "version": "0.0.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts"
+  },
+  "backstage": {
+    "role": "backend-plugin-module"
+  },
+  "scripts": {
+    "start": "backstage-cli package start",
+    "build": "backstage-cli package build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test",
+    "clean": "backstage-cli package clean",
+    "prepack": "backstage-cli package prepack",
+    "postpack": "backstage-cli package postpack"
+  },
+  "dependencies": {
+    "@backstage/backend-common": "workspace:^",
+    "@backstage/config": "workspace:^",
+    "@backstage/errors": "workspace:^",
+    "@backstage/integration": "workspace:^",
+    "@backstage/plugin-scaffolder-node": "workspace:^",
+    "@octokit/webhooks": "^10.0.0",
+    "libsodium-wrappers": "^0.7.11",
+    "octokit": "^2.0.0",
+    "octokit-plugin-create-pull-request": "^3.10.0",
+    "winston": "^3.2.1",
+    "yaml": "^2.0.0"
+  },
+  "devDependencies": {
+    "@backstage/backend-test-utils": "workspace:^",
+    "@backstage/cli": "workspace:^",
+    "@types/libsodium-wrappers": "^0.7.10",
+    "fs-extra": "10.1.0",
+    "jest-when": "^3.1.0",
+    "jsonschema": "^1.2.6"
+  },
+  "files": [
+    "dist"
+  ]
+}
