+++ plugins/scaffolder-backend-module-github/src/actions/github.ts
@@ -14,39 +14,49 @@
  * limitations under the License.
  */
 
+import { Config } from '@backstage/config';
 import { InputError } from '@backstage/errors';
 import {
   GithubCredentialsProvider,
   ScmIntegrationRegistry,
 } from '@backstage/integration';
 import { Octokit } from 'octokit';
-import { createTemplateAction } from '@backstage/plugin-scaffolder-node';
-import { parseRepoUrl } from '../publish/util';
+import {
+  createTemplateAction,
+  parseRepoUrl,
+} from '@backstage/plugin-scaffolder-node';
 import {
   createGithubRepoWithCollaboratorsAndTopics,
   getOctokitOptions,
+  initRepoPushAndProtect,
 } from './helpers';
 import * as inputProps from './inputProperties';
 import * as outputProps from './outputProperties';
-import { examples } from './githubRepoCreate.examples';
+import { examples } from './github.examples';
 
 /**
- * Creates a new action that initializes a git repository
+ * Creates a new action that initializes a git repository of the content in the workspace
+ * and publishes it to GitHub.
  *
  * @public
  */
-export function createGithubRepoCreateAction(options: {
+export function createPublishGithubAction(options: {
   integrations: ScmIntegrationRegistry;
+  config: Config;
   githubCredentialsProvider?: GithubCredentialsProvider;
 }) {
-  const { integrations, githubCredentialsProvider } = options;
+  const { integrations, config, githubCredentialsProvider } = options;
 
   return createTemplateAction<{
     repoUrl: string;
     description?: string;
     homepage?: string;
     access?: string;
+    defaultBranch?: string;
+    protectDefaultBranch?: boolean;
+    protectEnforceAdmins?: boolean;
     deleteBranchOnMerge?: boolean;
+    gitCommitMessage?: string;
     gitAuthorName?: string;
     gitAuthorEmail?: string;
     allowRebaseMerge?: boolean;
@@ -55,18 +65,24 @@ export function createGithubRepoCreateAction(options: {
     squashMergeCommitMessage?: 'PR_BODY' | 'COMMIT_MESSAGES' | 'BLANK';
     allowMergeCommit?: boolean;
     allowAutoMerge?: boolean;
-    requireCodeOwnerReviews?: boolean;
-    bypassPullRequestAllowances?: {
-      users?: string[];
-      teams?: string[];
-      apps?: string[];
-    };
+    sourcePath?: string;
+    bypassPullRequestAllowances?:
+      | {
+          users?: string[];
+          teams?: string[];
+          apps?: string[];
+        }
+      | undefined;
     requiredApprovingReviewCount?: number;
-    restrictions?: {
-      users: string[];
-      teams: string[];
-      apps?: string[];
-    };
+    restrictions?:
+      | {
+          users: string[];
+          teams: string[];
+          apps?: string[];
+        }
+      | undefined;
+    requireCodeOwnerReviews?: boolean;
+    dismissStaleReviews?: boolean;
     requiredStatusCheckContexts?: string[];
     requireBranchesToBeUpToDate?: boolean;
     requiredConversationResolution?: boolean;
@@ -86,17 +102,18 @@ export function createGithubRepoCreateAction(options: {
           access: 'pull' | 'push' | 'admin' | 'maintain' | 'triage';
         }
     >;
-    hasProjects?: boolean;
-    hasWiki?: boolean;
-    hasIssues?: boolean;
+    hasProjects?: boolean | undefined;
+    hasWiki?: boolean | undefined;
+    hasIssues?: boolean | undefined;
     token?: string;
     topics?: string[];
     repoVariables?: { [key: string]: string };
     secrets?: { [key: string]: string };
-    requireCommitSigning?: boolean;
+    requiredCommitSigning?: boolean;
   }>({
-    id: 'github:repo:create',
-    description: 'Creates a GitHub repository.',
+    id: 'publish:github',
+    description:
+      'Initializes a git repository of contents in workspace and publishes it to GitHub.',
     examples,
     schema: {
       input: {
@@ -107,22 +124,30 @@ export function createGithubRepoCreateAction(options: {
           description: inputProps.description,
           homepage: inputProps.homepage,
           access: inputProps.access,
-          requireCodeOwnerReviews: inputProps.requireCodeOwnerReviews,
           bypassPullRequestAllowances: inputProps.bypassPullRequestAllowances,
           requiredApprovingReviewCount: inputProps.requiredApprovingReviewCount,
           restrictions: inputProps.restrictions,
+          requireCodeOwnerReviews: inputProps.requireCodeOwnerReviews,
+          dismissStaleReviews: inputProps.dismissStaleReviews,
           requiredStatusCheckContexts: inputProps.requiredStatusCheckContexts,
           requireBranchesToBeUpToDate: inputProps.requireBranchesToBeUpToDate,
           requiredConversationResolution:
             inputProps.requiredConversationResolution,
           repoVisibility: inputProps.repoVisibility,
+          defaultBranch: inputProps.defaultBranch,
+          protectDefaultBranch: inputProps.protectDefaultBranch,
+          protectEnforceAdmins: inputProps.protectEnforceAdmins,
           deleteBranchOnMerge: inputProps.deleteBranchOnMerge,
+          gitCommitMessage: inputProps.gitCommitMessage,
+          gitAuthorName: inputProps.gitAuthorName,
+          gitAuthorEmail: inputProps.gitAuthorEmail,
           allowMergeCommit: inputProps.allowMergeCommit,
           allowSquashMerge: inputProps.allowSquashMerge,
           squashMergeCommitTitle: inputProps.squashMergeCommitTitle,
           squashMergeCommitMessage: inputProps.squashMergeCommitMessage,
           allowRebaseMerge: inputProps.allowRebaseMerge,
           allowAutoMerge: inputProps.allowAutoMerge,
+          sourcePath: inputProps.sourcePath,
           collaborators: inputProps.collaborators,
           hasProjects: inputProps.hasProjects,
           hasWiki: inputProps.hasWiki,
@@ -139,6 +164,7 @@ export function createGithubRepoCreateAction(options: {
         properties: {
           remoteUrl: outputProps.remoteUrl,
           repoContentsUrl: outputProps.repoContentsUrl,
+          commitHash: outputProps.commitHash,
         },
       },
     },
@@ -148,8 +174,22 @@ export function createGithubRepoCreateAction(options: {
         description,
         homepage,
         access,
+        requireCodeOwnerReviews = false,
+        dismissStaleReviews = false,
+        bypassPullRequestAllowances,
+        requiredApprovingReviewCount = 1,
+        restrictions,
+        requiredStatusCheckContexts = [],
+        requireBranchesToBeUpToDate = true,
+        requiredConversationResolution = false,
         repoVisibility = 'private',
+        defaultBranch = 'master',
+        protectDefaultBranch = true,
+        protectEnforceAdmins = true,
         deleteBranchOnMerge = false,
+        gitCommitMessage = 'initial commit',
+        gitAuthorName,
+        gitAuthorEmail,
         allowMergeCommit = true,
         allowSquashMerge = true,
         squashMergeCommitTitle = 'COMMIT_OR_PR_TITLE',
@@ -164,6 +204,7 @@ export function createGithubRepoCreateAction(options: {
         repoVariables,
         secrets,
         token: providedToken,
+        requiredCommitSigning = false,
       } = ctx.input;
 
       const octokitOptions = await getOctokitOptions({
@@ -205,7 +246,39 @@ export function createGithubRepoCreateAction(options: {
         ctx.logger,
       );
 
-      ctx.output('remoteUrl', newRepo.clone_url);
+      const remoteUrl = newRepo.clone_url;
+      const repoContentsUrl = `${newRepo.html_url}/blob/${defaultBranch}`;
+
+      const commitResult = await initRepoPushAndProtect(
+        remoteUrl,
+        octokitOptions.auth,
+        ctx.workspacePath,
+        ctx.input.sourcePath,
+        defaultBranch,
+        protectDefaultBranch,
+        protectEnforceAdmins,
+        owner,
+        client,
+        repo,
+        requireCodeOwnerReviews,
+        bypassPullRequestAllowances,
+        requiredApprovingReviewCount,
+        restrictions,
+        requiredStatusCheckContexts,
+        requireBranchesToBeUpToDate,
+        requiredConversationResolution,
+        config,
+        ctx.logger,
+        gitCommitMessage,
+        gitAuthorName,
+        gitAuthorEmail,
+        dismissStaleReviews,
+        requiredCommitSigning,
+      );
+
+      ctx.output('commitHash', commitResult?.commitHash);
+      ctx.output('remoteUrl', remoteUrl);
+      ctx.output('repoContentsUrl', repoContentsUrl);
     },
   });
 }
