+++ plugins/scaffolder-backend-module-github/src/actions/githubIssuesLabel.ts
@@ -18,11 +18,13 @@ import {
   GithubCredentialsProvider,
   ScmIntegrationRegistry,
 } from '@backstage/integration';
-import { createTemplateAction } from '@backstage/plugin-scaffolder-node';
+import {
+  createTemplateAction,
+  parseRepoUrl,
+} from '@backstage/plugin-scaffolder-node';
 import { assertError, InputError } from '@backstage/errors';
 import { Octokit } from 'octokit';
 import { getOctokitOptions } from './helpers';
-import { parseRepoUrl } from '../publish/util';
 import { examples } from './githubIssuesLabel.examples';
 
 /**
