+++ plugins/scaffolder-backend-module-github/src/actions/githubRepoCreate.ts
@@ -14,47 +14,41 @@
  * limitations under the License.
  */
 
-import { Config } from '@backstage/config';
 import { InputError } from '@backstage/errors';
 import {
   GithubCredentialsProvider,
   ScmIntegrationRegistry,
 } from '@backstage/integration';
 import { Octokit } from 'octokit';
-import { createTemplateAction } from '@backstage/plugin-scaffolder-node';
+import {
+  createTemplateAction,
+  parseRepoUrl,
+} from '@backstage/plugin-scaffolder-node';
 import {
   createGithubRepoWithCollaboratorsAndTopics,
   getOctokitOptions,
-  initRepoPushAndProtect,
-} from '../github/helpers';
-import * as inputProps from '../github/inputProperties';
-import * as outputProps from '../github/outputProperties';
-import { parseRepoUrl } from './util';
-import { examples } from './github.examples';
+} from './helpers';
+import * as inputProps from './inputProperties';
+import * as outputProps from './outputProperties';
+import { examples } from './githubRepoCreate.examples';
 
 /**
- * Creates a new action that initializes a git repository of the content in the workspace
- * and publishes it to GitHub.
+ * Creates a new action that initializes a git repository
  *
  * @public
  */
-export function createPublishGithubAction(options: {
+export function createGithubRepoCreateAction(options: {
   integrations: ScmIntegrationRegistry;
-  config: Config;
   githubCredentialsProvider?: GithubCredentialsProvider;
 }) {
-  const { integrations, config, githubCredentialsProvider } = options;
+  const { integrations, githubCredentialsProvider } = options;
 
   return createTemplateAction<{
     repoUrl: string;
     description?: string;
     homepage?: string;
     access?: string;
-    defaultBranch?: string;
-    protectDefaultBranch?: boolean;
-    protectEnforceAdmins?: boolean;
     deleteBranchOnMerge?: boolean;
-    gitCommitMessage?: string;
     gitAuthorName?: string;
     gitAuthorEmail?: string;
     allowRebaseMerge?: boolean;
@@ -63,24 +57,18 @@ export function createPublishGithubAction(options: {
     squashMergeCommitMessage?: 'PR_BODY' | 'COMMIT_MESSAGES' | 'BLANK';
     allowMergeCommit?: boolean;
     allowAutoMerge?: boolean;
-    sourcePath?: string;
-    bypassPullRequestAllowances?:
-      | {
-          users?: string[];
-          teams?: string[];
-          apps?: string[];
-        }
-      | undefined;
-    requiredApprovingReviewCount?: number;
-    restrictions?:
-      | {
-          users: string[];
-          teams: string[];
-          apps?: string[];
-        }
-      | undefined;
     requireCodeOwnerReviews?: boolean;
-    dismissStaleReviews?: boolean;
+    bypassPullRequestAllowances?: {
+      users?: string[];
+      teams?: string[];
+      apps?: string[];
+    };
+    requiredApprovingReviewCount?: number;
+    restrictions?: {
+      users: string[];
+      teams: string[];
+      apps?: string[];
+    };
     requiredStatusCheckContexts?: string[];
     requireBranchesToBeUpToDate?: boolean;
     requiredConversationResolution?: boolean;
@@ -100,18 +88,17 @@ export function createPublishGithubAction(options: {
           access: 'pull' | 'push' | 'admin' | 'maintain' | 'triage';
         }
     >;
-    hasProjects?: boolean | undefined;
-    hasWiki?: boolean | undefined;
-    hasIssues?: boolean | undefined;
+    hasProjects?: boolean;
+    hasWiki?: boolean;
+    hasIssues?: boolean;
     token?: string;
     topics?: string[];
     repoVariables?: { [key: string]: string };
     secrets?: { [key: string]: string };
-    requiredCommitSigning?: boolean;
+    requireCommitSigning?: boolean;
   }>({
-    id: 'publish:github',
-    description:
-      'Initializes a git repository of contents in workspace and publishes it to GitHub.',
+    id: 'github:repo:create',
+    description: 'Creates a GitHub repository.',
     examples,
     schema: {
       input: {
@@ -122,30 +109,22 @@ export function createPublishGithubAction(options: {
           description: inputProps.description,
           homepage: inputProps.homepage,
           access: inputProps.access,
+          requireCodeOwnerReviews: inputProps.requireCodeOwnerReviews,
           bypassPullRequestAllowances: inputProps.bypassPullRequestAllowances,
           requiredApprovingReviewCount: inputProps.requiredApprovingReviewCount,
           restrictions: inputProps.restrictions,
-          requireCodeOwnerReviews: inputProps.requireCodeOwnerReviews,
-          dismissStaleReviews: inputProps.dismissStaleReviews,
           requiredStatusCheckContexts: inputProps.requiredStatusCheckContexts,
           requireBranchesToBeUpToDate: inputProps.requireBranchesToBeUpToDate,
           requiredConversationResolution:
             inputProps.requiredConversationResolution,
           repoVisibility: inputProps.repoVisibility,
-          defaultBranch: inputProps.defaultBranch,
-          protectDefaultBranch: inputProps.protectDefaultBranch,
-          protectEnforceAdmins: inputProps.protectEnforceAdmins,
           deleteBranchOnMerge: inputProps.deleteBranchOnMerge,
-          gitCommitMessage: inputProps.gitCommitMessage,
-          gitAuthorName: inputProps.gitAuthorName,
-          gitAuthorEmail: inputProps.gitAuthorEmail,
           allowMergeCommit: inputProps.allowMergeCommit,
           allowSquashMerge: inputProps.allowSquashMerge,
           squashMergeCommitTitle: inputProps.squashMergeCommitTitle,
           squashMergeCommitMessage: inputProps.squashMergeCommitMessage,
           allowRebaseMerge: inputProps.allowRebaseMerge,
           allowAutoMerge: inputProps.allowAutoMerge,
-          sourcePath: inputProps.sourcePath,
           collaborators: inputProps.collaborators,
           hasProjects: inputProps.hasProjects,
           hasWiki: inputProps.hasWiki,
@@ -162,7 +141,6 @@ export function createPublishGithubAction(options: {
         properties: {
           remoteUrl: outputProps.remoteUrl,
           repoContentsUrl: outputProps.repoContentsUrl,
-          commitHash: outputProps.commitHash,
         },
       },
     },
@@ -172,22 +150,8 @@ export function createPublishGithubAction(options: {
         description,
         homepage,
         access,
-        requireCodeOwnerReviews = false,
-        dismissStaleReviews = false,
-        bypassPullRequestAllowances,
-        requiredApprovingReviewCount = 1,
-        restrictions,
-        requiredStatusCheckContexts = [],
-        requireBranchesToBeUpToDate = true,
-        requiredConversationResolution = false,
         repoVisibility = 'private',
-        defaultBranch = 'master',
-        protectDefaultBranch = true,
-        protectEnforceAdmins = true,
         deleteBranchOnMerge = false,
-        gitCommitMessage = 'initial commit',
-        gitAuthorName,
-        gitAuthorEmail,
         allowMergeCommit = true,
         allowSquashMerge = true,
         squashMergeCommitTitle = 'COMMIT_OR_PR_TITLE',
@@ -202,7 +166,6 @@ export function createPublishGithubAction(options: {
         repoVariables,
         secrets,
         token: providedToken,
-        requiredCommitSigning = false,
       } = ctx.input;
 
       const octokitOptions = await getOctokitOptions({
@@ -244,39 +207,7 @@ export function createPublishGithubAction(options: {
         ctx.logger,
       );
 
-      const remoteUrl = newRepo.clone_url;
-      const repoContentsUrl = `${newRepo.html_url}/blob/${defaultBranch}`;
-
-      const commitResult = await initRepoPushAndProtect(
-        remoteUrl,
-        octokitOptions.auth,
-        ctx.workspacePath,
-        ctx.input.sourcePath,
-        defaultBranch,
-        protectDefaultBranch,
-        protectEnforceAdmins,
-        owner,
-        client,
-        repo,
-        requireCodeOwnerReviews,
-        bypassPullRequestAllowances,
-        requiredApprovingReviewCount,
-        restrictions,
-        requiredStatusCheckContexts,
-        requireBranchesToBeUpToDate,
-        requiredConversationResolution,
-        config,
-        ctx.logger,
-        gitCommitMessage,
-        gitAuthorName,
-        gitAuthorEmail,
-        dismissStaleReviews,
-        requiredCommitSigning,
-      );
-
-      ctx.output('commitHash', commitResult?.commitHash);
-      ctx.output('remoteUrl', remoteUrl);
-      ctx.output('repoContentsUrl', repoContentsUrl);
+      ctx.output('remoteUrl', newRepo.clone_url);
     },
   });
 }
