+++ plugins/scaffolder-backend-module-github/src/actions/index.ts
@@ -21,3 +21,6 @@ export { createGithubRepoPushAction } from './githubRepoPush';
 export { createGithubWebhookAction } from './githubWebhook';
 export { createGithubDeployKeyAction } from './githubDeployKey';
 export { createGithubEnvironmentAction } from './githubEnvironment';
+
+export { createPublishGithubPullRequestAction } from './githubPullRequest';
+export { createPublishGithubAction } from './github';
