+++ plugins/scaffolder-backend-module-github/src/actions/outputProperties.ts
